move decoder to create package
simplifies code
